requirements.txt
blinker==1.8.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
mysqlclient==2.2.4
packaging==24.1
psycopg2==2.9.9
pycparser==2.22
PyMySQL==1.1.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
